feat(APP): 增加无人机悬停功能并优化飞行控制逻辑
- 新增 Drone_Idle 状态，用于起飞前的准备阶段
- 实现高度和速度的 PID 控制，提高悬停稳定性
- 优化遥控器信号丢失时的故障处理逻辑
- 调整 IMU 校准时间和方法，提高传感器精度
- 重构部分代码，提高可读性和可维护性

+++ Common/IMU/Com_IMU.c
@@ -1,6 +1,8 @@
 #include "Com_IMU.h"
 #include "math.h"
 
+
+
 /* ============================欧拉角计算================================== */
 /* ===============================开始===================================== */
 
@@ -32,7 +34,7 @@ static float Q_rsqrt(float number)
     return y;
 }
 
-static double normAccz; /* z轴上的加速度 */
+static float normAccz; /* z轴上的加速度 */
 /**
  * @description: 根据mpu的6轴数据, 获取表征姿态的欧拉角
  * @param {GyroAccel_Struct} *gyroAccel mpu的6轴数据
@@ -124,6 +126,7 @@ void Com_IMU_GetEulerAngle(GyroAccel_Struct *gyroAccel,
  * @description: 获取Z轴上的加速度 (如果已经倾斜,会考虑z轴上加速度的合成)
  * @return {*}
  */
+
 float Com_IMU_GetNormAccZ(void)
 {
     return normAccz;
